correction of the constructor

+++ CPP02/ex02/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: aharder <aharder@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/01 13:02:48 by aharder           #+#    #+#             */
-/*   Updated: 2025/10/28 17:21:05 by aharder          ###   ########.fr       */
+/*   Updated: 2025/10/28 17:24:10 by aharder          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ Fixed::Fixed(float const raw)
 Fixed::Fixed(int const raw)
 {
 	std::cout << "Int constructor called" << std::endl;
-	Fixed::value = round(raw * (pow(2, this->fract_bits)));
+	Fixed::value = roundf(raw * (pow(2, this->fract_bits)));
     std::cout << "[DEBUG] " << Fixed::toInt() << std::endl;
 }
